RR 17/08/2020
The LPC1769 allows functions to be multiplexed to various pins.
Board manufacturers choose which pins are used for which function.

This allows to choose any pin combination for UARTs to be defined by additional optional symbols.

It can resolve the BigTreeTech used different pins for UART3, compared to other manufaturers.

In this case defineing UART_PIN_82_85 enables UART3, which is available at the Wifi port.

The symbols per UART are:

		 For UART0: There is only one set of pins: pin 98 and 99, so no define needed

		 For UART1: 			        default is pin 62 and 63
					define UART1_PIN_75_74 to use  pin 75 and 74

		 For UART2: 			        default is pin 48 and 49
					define UART2_PIN_65_64 to use  pin 65 and 44

		 For UART3: 			        default is pin 46 and 47
					define UART3_PIN_7_6   to use  pin  7 and  6
					define UART3_PIN_82_85 to use  pin 82 and 85
DISCLASIMER

Currently I have *only* tested correct functioning of this setting for UART3 with UART3_PIN_82_85 for my BTT SKR 1.4 Turbo.
All settings were derived from the hardware documentation:
https://github.com/bigtreetech/BIGTREETECH-SKR-V1.3/blob/master/BTT%20SKR%20V1.4/Datasheet/LPC17-32-IC.pdf